// sgm/src/app/pages/pacientes/interfaces/paciente.interface.ts
export interface Paciente{
  id: String;
  numdocpaciente: number;
  pnombrepaciente: String;
  snombrepaciente: String;
  papellidopaciente: String;
  sapellidopaciente: String;
  tipodocpaciente: String;
  birthpaciente: String;
  generopaciente: String;
  paispaciente: String;
  movilpaciente: number;
  telfijopaciente: number;
  dirpaciente: String;
  ciudadpaciente: String;
  sedepaciente: String;
  emailpaciente: String;
  estadopaciente: Boolean;
  recibircomunicados: Boolean;
}